refactor(pilaf): rename part 2
